+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add Vserver resources diagram showing SVM information, allocated aggregates, and network interfaces (LIFs)
+- Add Network ports diagram showing all network ports and their associated LIFs
 - Add Volume resources usage to the vserver diagram
 - Add node owner (Home Node) to network interface tables (Vserver, Cluster, Management, Intercluster, and Data network interfaces)
 - Add Health Check best practices for Network Broadcast Domains